Controller Added
I did a little bit of housekeeping and added in a few things here in
there to help implement the coding for the controller. Here is a short
list of what I did:
- Removed all all array variables; they were causing problems and the
arithmetic was going to be an issue.
- Added in float variables to compensate for the lack of array
variables.
- Changed the names of one or two variables to have the code be less
confusing to work with.
- Added in new categories of variables to make it more organized and
less cluttered.
- Added in two new functions to help implement the controller code:
sign() and fix()
- Added in a steering controller.

Please let me know if you catch any big errors or problems and I will
make the changes as necessary.