Outline of Starting Position Code
I have added in what I assume is going to be the way we implement the
starting position code. I have essentially done most of the legwork in
getting it written into an relatively understandable format. All that
needs to really be done on your end is get the circuitry wired together
so that the button can be implemented as is expected in the code and the
potentiometer and pretty accurately and reliably be used to set the
starting position.

If there are any problems let me know and I'll make sure to make any
changes ASAP