Code Fix
A lot of errors were coming up when compiling and decompiling a lot of
stuff. I rearranged some stuff (which to be honest may not have done
anything), and changed some syntax stuff. I really just verified the
code without the potentiometer part until the code compiled without any
errors. In the current form of the code it should do that.

Important to note that I did change the name for servoBearing back to
servoBearing to avoid a redefinition error but then commented out the
section of the code that uses that variable. This was to avoid any weird
definition errors from popping up.  servoBearing should be defined as
phi_desired and only defined once (it currently would be defined in both
the Target Variables section as well as the servo variables section).
The old servo code also needs to be reworked or changed altogether since
it will cause a conflicting error in the code.

In addition, the code relies on the presumption that the variables
startingPosX and startingPosY have been defined in the initial setup
loop (I have already defined them to test for that error). This means
that they just need to be defined in the setup() loop with the
potentiometer function.