CONTROLLER WORKS: PART I
I got the controller to work.

- I changed the thetaOffset to point a certain direction in my room, can
easily be changed back to what it needs to be.
- I added a part in the code where the servo points itself forwards at
the beginning. Just a little quality of life thing.
- I was messing with some Serial.print stuff for troubleshooting so just
kinda ignore a lot of the random Serial.print stuff.
- I changed the amout of time before sleep because I couldn't do
everything I wanted within the time it was giving me. Easily reversible.
- I actually added in the steering() controller into the program
- I changed my fix() function since its now supposed to fix angles from
degrees (it was originally fixing angles in radians).
- I was messing with some of the initializations for the desired
positions and corrected some stuff. I also fixed some of the function
calculations since they were calculating stuff entirely incorrectly.
- The way the controller was originally set up it used a bunch of
position errors and orientation errors to decide where to set the servo.
However it doesn't set the angle based on the way the servoMagnetometer
code sets it. The code from the servoMag was also originally incorrect
in that the 90-(theta-servoBearing) part was wrong. I corrected it to
what it was supposed to be and it works as expected. It wasn't quite the
same here in the controller code, but now the servo DOES try to point
towards its target.
- I initialized one or two extra variables for housekeeping of my
variables, so don't be shocked to find one or two laying around.
- I also put in the correction and limiting function lines.
- A good number of things that look like they used to be function parts
of the code have been commented out, so don't be surprised by that. That
was all on purpose because there was probably something faulty with it
that I still need to figure out.

TESTING
- I tested the code from two positions (B2 and D4). Each position was on
a different side of the x-axis and had a different starting orientation.
- I tested the code up until the conditional kicked in and the steering
law changed.
- It worked relatively well with the first position (B2), but was a
little frustrating to get right with the second position (D4). I needed
to maybe rework the conditionals a little more and actually test the
robot with the steering attached.
- That all being said though, it actually works: (1) The robot attempts
to face the steering forwards until it hits the channel, then; (2) The
target changes to lead the robot up the channel.